+++ src/components/Modals/AproveReceiptsModal.tsx
@@ -1,0 +1,75 @@
+/**
+ * AproveReceiptsModal component handles the approval of receipts 
+ * in the accounts payable module.
+ */
+
+import { useCallback } from "react";
+import { apiRequest } from "@utils/apiClient";
+import ModalWrapper from "@/components/Modals/ModalWrapper";
+import Button from "@/components/Buttons/Button";
+
+interface Props {
+  receipt_id: number;
+  title: string;
+  message: string;
+  redirection: string;
+  modal_type: "success" | "warning";
+  variant?: "filled" | "border" | "empty";
+  children: React.ReactNode;
+  disabled?: boolean;
+  token: string;
+}
+
+export default function ApproveReceiptStatus({
+  receipt_id,
+  title,
+  message,
+  redirection,
+  modal_type,
+  variant,
+  children,
+  disabled = false,
+  token
+}: Props) {
+  const handleConfirm = useCallback(async () => {
+    try {
+        const url = `/accounts-payable/validate-receipt/${receipt_id}`;
+      await apiRequest(url, { 
+        method: "PUT", 
+        data: {"approval": 1},
+        headers: { Authorization: `Bearer ${token}` }
+      });
+      alert(`Aprobado correctamente`)
+
+      if (redirection) {
+        window.location.reload();
+      } else {
+        window.location.reload();
+      }
+    } catch (error) {
+      console.error("Error en la solicitud:", error);
+    }
+  }, [receipt_id, redirection]);
+
+  return (
+    <ModalWrapper
+      title={title}
+      message={message}
+      color="success"
+      modal_type={modal_type}
+      onConfirm={handleConfirm}
+      disabled={disabled}
+      triggerElement={
+        <Button
+          color="success"
+          variant="filled"
+          size="medium"
+          disabled={disabled}
+          customSizeClass="w-full"
+        >
+          {children}
+        </Button>
+      }
+    />
+  );
+}