+++ src/components/Table/TableRow.tsx
@@ -2,7 +2,7 @@
  * Table row
  **/
 
-import Button from '@components/Button';
+import Button from '@/components/Buttons/Button';
 
 interface Column {
   key: string;